# static/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Food Donation API",
        "version": "1.0.0"
    },
    "paths": {
        "/register": {
            "post": {
                "summary": "Register a new user",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "username": {
                                    "type": "string"
                                },
                                "password": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User created"
                    }
                }
            }
        },
        "/login": {
            "post": {
                "summary": "Login a user",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "username": {
                                    "type": "string"
                                },
                                "password": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Access token returned"
                    }
                }
            }
        },
        "/donations": {
            "post": {
                "summary": "Create a new donation",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "food_item": {
                                    "type": "string"
                                },
                                "quantity": {
                                    "type": "integer"
                                },
                                "location": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Donation created"
                    }
                }
            },
            "get": {
                "summary": "Get all donations",
                "responses": {
                    "200": {
                        "description": "List of donations"
                    }
                }
            }
        }
    }
}